implemented short forecast, next taks is to show progress bar when they are still loading

+++ app/src/main/java/com/bignerdranch/android/weather/core/data/dto/ShortForecastDto.kt
@@ -11,7 +11,7 @@ fun ShortForecastDto.toShortForecast(): ShortForecast =
 
 fun ForecastDayDto.toForecastDay(): ForecastDay =
     ForecastDay(
-        day = this.date.substring(this.date.length-2).toInt(),
+        day = this.date.substring(this.date.length-2),
         maxTempInCelsius = this.day.maxTempInCelsius,
         minTempInCelsius = day.minTempInCelsius,
         maxTempInFahrenheit = day.maxTempInFahrenheit,
